+++ src/main/java/va/vt/cbil/LeftGroupPanel.java
@@ -1209,7 +1209,7 @@ class LeftGroupPanel {
 					case 4:
 						int ignoreMerge = jTF51.isSelected()?1:0;
 						int mergeEventDiscon = Integer.parseInt(jTF52.getText());
-						int mergeEventCorr = Integer.parseInt(jTF53.getText());
+						float mergeEventCorr = Float.parseFloat(jTF53.getText());
 						int mergeEventMaxTimeDif = Integer.parseInt(jTF54.getText());
 						imageDealer.setStep5(ignoreMerge, mergeEventDiscon, mergeEventCorr,mergeEventMaxTimeDif);
 						imageDealer.step5Start();
